GMMHMM Model Class  Built, but covariance matrix issue; not positivev semidefinite
